whoops, div is implemented to complement mod, and is jython floordiv function.
svn path=/autoplot/trunk/QDataSet/; revision=947

+++ QDataSet/src/org/virbo/dsops/Ops.java
@@ -543,7 +543,7 @@ public class Ops {
     public static QDataSet div( QDataSet ds1, QDataSet ds2 ) {
         return applyBinaryOp(ds1, ds2, new BinaryOp() {
             public double op(double d1, double d2) {
-                return (int)d1 / (int)d2;
+                return (int) ( d1 / d2 );
             }
         });
     }
